+++ src/Symfony/Component/Security/Core/Security.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Security\Core;
 
 use Psr\Container\ContainerInterface;
+use Symfony\Bundle\SecurityBundle\Security\Security as NewSecurityHelper;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
@@ -19,20 +20,47 @@ use Symfony\Component\Security\Core\User\UserInterface;
 /**
  * Helper class for commonly-needed security tasks.
  *
- * @final
+ * @deprecated since Symfony 6.2, use \Symfony\Bundle\SecurityBundle\Security\Security instead
  */
 class Security implements AuthorizationCheckerInterface
 {
+    /**
+     * @deprecated since Symfony 6.2, use \Symfony\Bundle\SecurityBundle\Security\Security::ACCESS_DENIED_ERROR instead
+     *
+     * In 7.0, move this constant to the NewSecurityHelper class and make it reference SecurityRequestAttributes:ACCESS_DENIED_ERROR.
+     */
     public const ACCESS_DENIED_ERROR = '_security.403_error';
+
+    /**
+     * @deprecated since Symfony 6.2, use \Symfony\Bundle\SecurityBundle\Security\Security::AUTHENTICATION_ERROR instead
+     *
+     * In 7.0, move this constant to the NewSecurityHelper class and make it reference SecurityRequestAttributes:AUTHENTICATION_ERROR.
+     */
     public const AUTHENTICATION_ERROR = '_security.last_error';
+
+    /**
+     * @deprecated since Symfony 6.2, use \Symfony\Bundle\SecurityBundle\Security\Security::LAST_USERNAME instead
+     *
+     * In 7.0, move this constant to the NewSecurityHelper class and make it reference SecurityRequestAttributes:LAST_USERNAME.
+     */
     public const LAST_USERNAME = '_security.last_username';
+
+    /**
+     * @deprecated since Symfony 6.2, use \Symfony\Component\Security\Http\Authenticator\AuthenticatorInterface::MAX_USERNAME_LENGTH instead
+     *
+     *  In 7.0, move this constant to the NewSecurityHelper class and make it reference AuthenticatorInterface:MAX_USERNAME_LENGTH.
+     */
     public const MAX_USERNAME_LENGTH = 4096;
 
     private ContainerInterface $container;
 
-    public function __construct(ContainerInterface $container)
+    public function __construct(ContainerInterface $container, bool $triggerDeprecation = true)
     {
         $this->container = $container;
+
+        if ($triggerDeprecation) {
+            trigger_deprecation('symfony/security-core', '6.2', 'The "%s" class is deprecated, use "%s" instead.', __CLASS__, NewSecurityHelper::class);
+        }
     }
 
     public function getUser(): ?UserInterface