+++ src/Symfony/Component/Security/Http/Authentication/AuthenticationUtils.php
@@ -14,7 +14,7 @@ namespace Symfony\Component\Security\Http\Authentication;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
-use Symfony\Component\Security\Core\Security;
+use Symfony\Component\Security\Http\SecurityRequestAttributes;
 
 /**
  * Extracts Security Errors from Request.
@@ -35,13 +35,13 @@ class AuthenticationUtils
         $request = $this->getRequest();
         $authenticationException = null;
 
-        if ($request->attributes->has(Security::AUTHENTICATION_ERROR)) {
-            $authenticationException = $request->attributes->get(Security::AUTHENTICATION_ERROR);
-        } elseif ($request->hasSession() && ($session = $request->getSession())->has(Security::AUTHENTICATION_ERROR)) {
-            $authenticationException = $session->get(Security::AUTHENTICATION_ERROR);
+        if ($request->attributes->has(SecurityRequestAttributes::AUTHENTICATION_ERROR)) {
+            $authenticationException = $request->attributes->get(SecurityRequestAttributes::AUTHENTICATION_ERROR);
+        } elseif ($request->hasSession() && ($session = $request->getSession())->has(SecurityRequestAttributes::AUTHENTICATION_ERROR)) {
+            $authenticationException = $session->get(SecurityRequestAttributes::AUTHENTICATION_ERROR);
 
             if ($clearSession) {
-                $session->remove(Security::AUTHENTICATION_ERROR);
+                $session->remove(SecurityRequestAttributes::AUTHENTICATION_ERROR);
             }
         }
 
@@ -52,11 +52,11 @@ class AuthenticationUtils
     {
         $request = $this->getRequest();
 
-        if ($request->attributes->has(Security::LAST_USERNAME)) {
-            return $request->attributes->get(Security::LAST_USERNAME, '');
+        if ($request->attributes->has(SecurityRequestAttributes::LAST_USERNAME)) {
+            return $request->attributes->get(SecurityRequestAttributes::LAST_USERNAME, '');
         }
 
-        return $request->hasSession() ? $request->getSession()->get(Security::LAST_USERNAME, '') : '';
+        return $request->hasSession() ? $request->getSession()->get(SecurityRequestAttributes::LAST_USERNAME, '') : '';
     }
 
     /**